pyproject.toml
[project]
name = "oc-da-python-projet-12"
version = "0.1.0"
description = ""
authors = [
    {name = "meridien22",email = "meridien.sig@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "pyjwt[crypto] (>=2.12.0,<3.0.0)",
    "ipython (>=9.11.0,<10.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "tabulate (>=0.10.0,<0.11.0)",
    "sentry-sdk (>=2.57.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-html (>=0.4.3,<0.5.0)"
]
